restaurant-recommender/backend/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.7
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
flask-cors==4.0.0